backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai>=1.50.0
python-dotenv==1.0.0
twilio==8.10.0
pyyaml==6.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL driver for Heroku
# pinecone-client removed - using OpenAI Assistants API instead
# tiktoken removed - no longer needed
